feat: add belongsToMany relation to group courses

+++ database/migrations/2024_09_08_212424_create_course_group_table.php
@@ -31,9 +31,11 @@ return new class extends Migration
         });
 
         // third, remove the old column
-        Schema::table('groups', function (Blueprint $table) {
-            $table->dropColumn('course_id');
-        });
+        if (!Schema::hasColumn('groups', 'course_id')) {
+            Schema::table('groups', function (Blueprint $table) {
+                $table->dropColumn('course_id');
+            });
+        }
     }
 
     /**
